Add button functionality and refactor code into appropiate files
