+++ test/sample-test.js
@@ -1,8 +1,11 @@
 import chai from 'chai';
 const expect = chai.expect;
 
+import { toggleLoginForm } from '../src/js/DOM.js';
+
 describe('See if the tests are running', function() {
   it('should return true', function() {
     expect(true).to.equal(true);
   });
 });
+
